rename: 디렉토리명 팀컨벤션에 맞도록 정정 #27
