+++ src/main/java/com/ktcloud/daangn/auth/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.ktcloud.daangn.auth.controller;
 
 import com.ktcloud.daangn.auth.dto.AuthSignupRequestDto;
 import com.ktcloud.daangn.auth.service.AuthService;
-import com.ktcloud.daangn.config.dto.BaseResponse;
+import com.ktcloud.daangn.common.dto.BaseResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
