+++ src/main/java/com/ktcloud/daangn/auth/dto/AuthSignupRequestDto.java
@@ -1,6 +1,6 @@
 package com.ktcloud.daangn.auth.dto;
 
-import com.ktcloud.daangn.config.valueObject.Address;
+import com.ktcloud.daangn.common.valueObject.Address;
 import com.ktcloud.daangn.member.entity.Member;
 import com.ktcloud.daangn.member.entity.MemberRole;
 import com.ktcloud.daangn.member.entity.ProviderToken;
