+++ src/main/java/com/ktcloud/daangn/auth/service/AuthServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ktcloud.daangn.auth.service;
 
 import com.ktcloud.daangn.auth.dto.AuthSignupRequestDto;
-import com.ktcloud.daangn.config.exception.InvalidInputException;
+import com.ktcloud.daangn.common.exception.InvalidInputException;
 import com.ktcloud.daangn.member.entity.Member;
 import com.ktcloud.daangn.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
