+++ src/main/java/com/ktcloud/daangn/common/ResultCode.java
@@ -1,4 +1,4 @@
-package com.ktcloud.daangn.config;
+package com.ktcloud.daangn.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
