+++ src/main/java/com/ktcloud/daangn/member/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.ktcloud.daangn.member.controller;
 
 import com.ktcloud.daangn.auth.dto.CustomUser;
-import com.ktcloud.daangn.config.dto.BaseResponse;
+import com.ktcloud.daangn.common.dto.BaseResponse;
 import com.ktcloud.daangn.member.dto.MemberInfoResponseDto;
 import com.ktcloud.daangn.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
