+++ src/main/java/com/ktcloud/daangn/member/service/MemberServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ktcloud.daangn.member.service;
 
-import com.ktcloud.daangn.config.exception.InvalidInputException;
+import com.ktcloud.daangn.common.exception.InvalidInputException;
 import com.ktcloud.daangn.member.dto.MemberInfoResponseDto;
 import com.ktcloud.daangn.member.entity.Member;
 import com.ktcloud.daangn.member.repository.MemberRepository;
