+++ src/test/java/com/ktcloud/daangn/notification/service/NotificationServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.ktcloud.daangn.notification.service;
 
-import com.ktcloud.daangn.config.exception.InvalidInputException;
+import com.ktcloud.daangn.common.exception.InvalidInputException;
 import com.ktcloud.daangn.member.entity.Member;
 import com.ktcloud.daangn.member.service.MemberService;
 import com.ktcloud.daangn.notification.dto.NotificationResponseDto;
